第九模块-项目实战vue-路飞学城开发/模块作业/LuffyCity/requirements.txt
Deprecated==1.2.13
Django==2.2
django-cors-headers==3.10.1
djangorestframework==3.8.2
importlib-metadata==4.8.3
packaging==21.3
Pillow==8.4.0
PyMySQL==1.0.2
pyparsing==3.0.7
pytz==2021.3
redis==4.1.4
sqlparse==0.4.2
typing_extensions==4.1.1
wrapt==1.13.3
zipp==3.6.0
